+++ src/output/types/fields/number-message-field.ts
@@ -68,6 +68,10 @@ export class NumberMessageField implements MessageField {
       printer.add(`if (instance.${this.attributeName} && instance.${this.attributeName}.length) {
         writer.writeRepeated${this.protoDataType}(${this.messageField.number}, instance.${this.attributeName});
       }`);
+    } else if (this.oneOf) {
+      printer.add(`if (instance.${this.attributeName} || instance.${this.attributeName} === 0) {
+        writer.write${this.protoDataType}(${this.messageField.number}, instance.${this.attributeName});
+      }`);
     } else {
       printer.add(`if (instance.${this.attributeName}) {
         writer.write${this.protoDataType}(${this.messageField.number}, instance.${this.attributeName});